dashboard/requirements.txt
openpyxl
pandas
redshift_connector
python-dotenv
xlrd
altair
streamlit

pyarrow
fastparquet
